Fix pgTAP storage test: cast file_size_limit comparison to bigint
The is() function requires matching types; file_size_limit is bigint
but the expected value was an untyped integer literal.

+++ supabase/tests/01_storage.sql
@@ -19,7 +19,7 @@ select is(
 -- -----------------------------------------------------------------------
 select is(
   (select file_size_limit from storage.buckets where id = 'post-images'),
-  1048576,
+  1048576::bigint,
   'post-images file_size_limit is 1MB'
 );
